Fix initial value of all store
With this change this `useStore(initialValue)` (or with  getStore / withStore) is valid. However, is doing an Object.assign with the parameter + the current store, this way only add the missing store variables and is not removing the existint ones.

````js
const { getStore } = createStore({ cart: { price: 0 }})
getStore({ username: 'Aral' })
// Now the store is { username: 'Aral', cart: { price: 0 } }
// Now the initial store is { username: 'Aral', cart: { price: 0 } }
```

The difference with setStore is that setStore is replacing all the store
without touching the initialStore:

```js
const { getStore } = createStore({ cart: { price: 0 }})
const [,setStore] = getStore()
setStore({ username: 'Aral' })
// Now the store is { username: 'Aral' }
// The initial store is the same that before: { cart: { price: 0 } }
```